Update 2024-04-09.md
I joined the narrative 4 story exchange. It’s really an amazing experience for me. I went through from to be excited, nervous and awkward, even I wanted to give up several times in the middle. But thankfully I didn’t. With the help of other story tellers, I made it. Not perfect, but I was so happy and satisfied with it. Building connections with people is always a good attempt. I really enjoyed it.